requirements.txt
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.3
pymongo==4.8.0
openai==1.59.6
langchain==0.3.7
langchain-core==0.3.15
langchain-openai==0.2.6
langgraph==0.2.35
fastapi==0.115.2
uvicorn==0.30.1
PyPDF2==3.0.1
langchain-openai==0.2.6